updated Like & Delete button styling

+++ src/components/DeleteArticle.jsx
@@ -7,20 +7,27 @@ import { db, storage } from '../firebaseConfig'
 const DeleteArticle = ({ id, imageUrl }) => {
 
     const handleDelete = async () => {
-        try {
-            await deleteDoc(doc(db, "Articles", id))
-            toast("Article successfully deleted", { type: "success" })
-            const storageRef = ref(storage, imageUrl)
-            await deleteObject(storageRef)
-        } catch (error) {
-            toast("Error deleting article", { type: "error" })
-            console.log(error)
+
+        if (window.confirm("Are you sure you want to delete this article?")) {
+            try {
+                await deleteDoc(doc(db, "Articles", id))
+                toast("Article successfully deleted", { type: "success" })
+                const storageRef = ref(storage, imageUrl)
+                await deleteObject(storageRef)
+            } catch (error) {
+                toast("Error deleting article", { type: "error" })
+                console.log(error)
+            }
         }
     }
 
     return (
         <div>
-            <button className="btn btn-danger" onClick={handleDelete}>Delete</button>
+            <i 
+                className="fa fa-times"
+                style={{ cursor: "pointer", color: "red" }}
+                onClick={handleDelete} 
+            />
         </div>
     )
 }
